requirements.txt
Flask==3.0.0
pandas==2.1.0
python-dateutil==2.8.2
gunicorn==21.2.0
